+++ .github/workflows/mypy.yml
@@ -5,10 +5,10 @@ on: [pull_request, push]
 jobs:
   mypy:
   runs_on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-python@v2\
-      - run: pip install mypy
-      - run: pip install -r requirements.txt
-      - run: mkdir --parents --verbose .mypy_cache
-      - run: mypy --ignore-missing-imports --install-types --non-interactive . || true
+  steps:
+    - uses: actions/checkout@v2
+    - uses: actions/setup-python@v2\
+    - run: pip install mypy
+    - run: pip install -r requirements.txt
+    - run: mkdir --parents --verbose .mypy_cache
+    - run: mypy --ignore-missing-imports --install-types --non-interactive . || true
